producer/requirements.txt
kafka-python
praw~=7.7.0
flask~=2.2.3
Faker~=17.6.0
yfinance~=0.2.12
mysql-connector-python~=8.0.32